third_party/blink/web_tests/external/wpt/css/css-text/animations/text-indent-composition-expected.txt

This is a testharness.js-based test.
Found 22 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Compositing CSS Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (-0.3) should be [20px hanging each-line]
  assert_equals: expected "20px hanging each - line " but got "345px "
[FAIL] Compositing CSS Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (0) should be [50px hanging each-line]
  assert_equals: expected "50px hanging each - line " but got "300px "
[FAIL] Compositing CSS Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (0.3) should be [80px hanging each-line]
  assert_equals: expected "80px hanging each - line " but got "255px "
[FAIL] Compositing CSS Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (0.6) should be [110px hanging each-line]
  assert_equals: expected "110px hanging each - line " but got "210px "
[FAIL] Compositing CSS Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (1) should be [150px hanging each-line]
  assert_equals: expected "150px hanging each - line " but got "150px "
[FAIL] Compositing CSS Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (1.5) should be [200px hanging each-line]
  assert_equals: expected "200px hanging each - line " but got "75px "
[FAIL] Compositing Web Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (-0.3) should be [20px hanging each-line]
  assert_equals: expected "20px hanging each - line " but got "345px "
[FAIL] Compositing Web Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (0) should be [50px hanging each-line]
  assert_equals: expected "50px hanging each - line " but got "300px "
[FAIL] Compositing Web Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (0.3) should be [80px hanging each-line]
  assert_equals: expected "80px hanging each - line " but got "255px "
[FAIL] Compositing Web Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (0.6) should be [110px hanging each-line]
  assert_equals: expected "110px hanging each - line " but got "210px "
[FAIL] Compositing Web Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (1) should be [150px hanging each-line]
  assert_equals: expected "150px hanging each - line " but got "150px "
[FAIL] Compositing Web Animations: property <text-indent> underlying [250px] from add [50px each-line hanging] to replace [150px hanging each-line] at (1.5) should be [200px hanging each-line]
  assert_equals: expected "200px hanging each - line " but got "75px "
[FAIL] Compositing CSS Animations: property <text-indent> underlying [250px each-line] from add [50px each-line] to replace [150px hanging] at (-0.3) should be [300px each-line]
  assert_equals: expected "300px each - line " but got "345px each - line "
[FAIL] Compositing CSS Animations: property <text-indent> underlying [250px each-line] from add [50px each-line] to replace [150px hanging] at (0.3) should be [300px each-line]
  assert_equals: expected "300px each - line " but got "255px each - line "
[FAIL] Compositing CSS Animations: property <text-indent> underlying [250px each-line] from add [50px each-line] to replace [150px hanging] at (0.6) should be [150px hanging]
  assert_equals: expected "150px hanging " but got "210px each - line "
[FAIL] Compositing CSS Animations: property <text-indent> underlying [250px each-line] from add [50px each-line] to replace [150px hanging] at (1) should be [150px hanging]
  assert_equals: expected "150px hanging " but got "150px each - line "
[FAIL] Compositing CSS Animations: property <text-indent> underlying [250px each-line] from add [50px each-line] to replace [150px hanging] at (1.5) should be [150px hanging]
  assert_equals: expected "150px hanging " but got "75px each - line "
[FAIL] Compositing Web Animations: property <text-indent> underlying [250px each-line] from add [50px each-line] to replace [150px hanging] at (-0.3) should be [300px each-line]
  assert_equals: expected "300px each - line " but got "345px each - line "
[FAIL] Compositing Web Animations: property <text-indent> underlying [250px each-line] from add [50px each-line] to replace [150px hanging] at (0.3) should be [300px each-line]
  assert_equals: expected "300px each - line " but got "255px each - line "
[FAIL] Compositing Web Animations: property <text-indent> underlying [250px each-line] from add [50px each-line] to replace [150px hanging] at (0.6) should be [150px hanging]
  assert_equals: expected "150px hanging " but got "210px each - line "
[FAIL] Compositing Web Animations: property <text-indent> underlying [250px each-line] from add [50px each-line] to replace [150px hanging] at (1) should be [150px hanging]
  assert_equals: expected "150px hanging " but got "150px each - line "
[FAIL] Compositing Web Animations: property <text-indent> underlying [250px each-line] from add [50px each-line] to replace [150px hanging] at (1.5) should be [150px hanging]
  assert_equals: expected "150px hanging " but got "75px each - line "
Harness: the test ran to completion.
